[chore][fileconsumer] Extract splitter, flusher, multiline, encoding into dedicated package
