+++ .chloggen/pkg-stanza-tokenize.yaml
@@ -7,19 +7,28 @@ change_type: deprecation
 component: pkg/stanza
 
 # A brief description of the change.  Surround your text with quotes ("") if it needs to start with a backtick (`).
-note: Deprecate encoding related elements of helper pacakge, in favor of new decoder package
+note: Deprecate tokenization related elements of helper pacakge, in favor of new tokenize package
 
 # Mandatory: One or more tracking issues related to the change. You can use the PR number here if no issue exists.
-issues: [26019]
+issues: [25914]
 
 # (Optional) One or more lines of additional information to render under the primary note.
 # These lines will be padded with 2 spaces and then inserted directly into the document.
 # Use pipe (|) for multiline entries.
 subtext: Includes the following deprecations |
-  - Decoder
-  - NewDecoder
-  - LookupEncoding
-  - IsNop
+  - Flusher
+  - FlusherConfig
+  - NewFlusherConfig
+  - Multiline
+  - MultilineConfig
+  - NewMultilineConfig
+  - NewLineStartSplitFunc
+  - NewLineEndSplitFunc
+  - NewNewlineSplitFunc
+  - Splitter
+  - SplitterConfig
+  - NewSplitterConfig
+  - SplitNone
 
 # If your change doesn't affect end users or the exported elements of any package,
 # you should instead start your pull request title with [chore] or use the "Skip Changelog" label.
