+++ pkg/stanza/fileconsumer/internal/splitter/multiline.go
@@ -7,16 +7,16 @@ import (
 	"bufio"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/decoder"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/tokenize"
 )
 
 type multilineFactory struct {
-	helper.SplitterConfig
+	tokenize.SplitterConfig
 }
 
 var _ Factory = (*multilineFactory)(nil)
 
-func NewMultilineFactory(splitter helper.SplitterConfig) Factory {
+func NewMultilineFactory(splitter tokenize.SplitterConfig) Factory {
 	return &multilineFactory{
 		SplitterConfig: splitter,
 	}
