+++ pkg/stanza/operator/helper/encoding.go
@@ -20,7 +20,6 @@ type EncodingConfig struct {
 // Deprecated: [v0.84.0] Use decoder.Decoder instead
 type Decoder = decoder.Decoder
 
-// Deprecated: [v0.84.0] Use decoder.New instead
 var NewDecoder = decoder.New
 
 // Deprecated: [v0.84.0] Use decoder.LookupEncoding instead
