+++ pkg/stanza/operator/input/syslog/config_test.go
@@ -9,10 +9,10 @@ import (
 
 	"go.opentelemetry.io/collector/config/configtls"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/tcp"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/udp"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/operatortest"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/tokenize"
 )
 
 func TestUnmarshal(t *testing.T) {
@@ -38,7 +38,7 @@ func TestUnmarshal(t *testing.T) {
 					cfg.TCP.ListenAddress = "10.0.0.1:9000"
 					cfg.TCP.AddAttributes = true
 					cfg.TCP.Encoding = "utf-16"
-					cfg.TCP.Multiline = helper.NewMultilineConfig()
+					cfg.TCP.Multiline = tokenize.NewMultilineConfig()
 					cfg.TCP.Multiline.LineStartPattern = "ABC"
 					cfg.TCP.TLS = &configtls.TLSServerSetting{
 						TLSSetting: configtls.TLSSetting{
@@ -62,7 +62,7 @@ func TestUnmarshal(t *testing.T) {
 					cfg.UDP.ListenAddress = "10.0.0.1:9000"
 					cfg.UDP.AddAttributes = true
 					cfg.UDP.Encoding = "utf-16"
-					cfg.UDP.Multiline = helper.NewMultilineConfig()
+					cfg.UDP.Multiline = tokenize.NewMultilineConfig()
 					cfg.UDP.Multiline.LineStartPattern = "ABC"
 					return cfg
 				}(),
