+++ pkg/stanza/operator/input/syslog/syslog_test.go
@@ -254,3 +254,5 @@ func TestOctetFramingSplitFunc(t *testing.T) {
 		t.Run(tc.Name, tc.RunFunc(splitFunc))
 	}
 }
+
+// TODO refactor test dependency away from internal?
