+++ pkg/stanza/operator/input/tcp/config_test.go
@@ -9,8 +9,8 @@ import (
 
 	"go.opentelemetry.io/collector/config/configtls"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/operatortest"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/tokenize"
 )
 
 func TestUnmarshal(t *testing.T) {
@@ -32,7 +32,7 @@ func TestUnmarshal(t *testing.T) {
 					cfg.ListenAddress = "10.0.0.1:9000"
 					cfg.AddAttributes = true
 					cfg.Encoding = "utf-8"
-					cfg.Multiline = helper.NewMultilineConfig()
+					cfg.Multiline = tokenize.NewMultilineConfig()
 					cfg.Multiline.LineStartPattern = "ABC"
 					cfg.TLS = &configtls.TLSServerSetting{
 						TLSSetting: configtls.TLSSetting{
