+++ pkg/stanza/operator/input/tcp/tcp.go
@@ -24,6 +24,7 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/decoder"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/tokenize"
 )
 
 const (
@@ -53,7 +54,7 @@ func NewConfigWithID(operatorID string) *Config {
 		InputConfig: helper.NewInputConfig(operatorID, operatorType),
 		BaseConfig: BaseConfig{
 			OneLogPerPacket: false,
-			Multiline:       helper.NewMultilineConfig(),
+			Multiline:       tokenize.NewMultilineConfig(),
 			Encoding:        "utf-8",
 		},
 	}
@@ -73,7 +74,7 @@ type BaseConfig struct {
 	AddAttributes               bool                        `mapstructure:"add_attributes,omitempty"`
 	OneLogPerPacket             bool                        `mapstructure:"one_log_per_packet,omitempty"`
 	Encoding                    string                      `mapstructure:"encoding,omitempty"`
-	Multiline                   helper.MultilineConfig      `mapstructure:"multiline,omitempty"`
+	Multiline                   tokenize.MultilineConfig    `mapstructure:"multiline,omitempty"`
 	PreserveLeadingWhitespaces  bool                        `mapstructure:"preserve_leading_whitespaces,omitempty"`
 	PreserveTrailingWhitespaces bool                        `mapstructure:"preserve_trailing_whitespaces,omitempty"`
 	MultiLineBuilder            MultiLineBuilderFunc
